requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
black==20.8b1
boto3==1.16.59
botocore==1.19.59
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
defusedxml==0.6.0
fastapi==0.63.0
fastapi-jwt-auth==0.5.0
greenlet==0.4.17
h11==0.11.0
httptools==0.1.1
idna==2.10
iniconfig==1.1.1
isort==5.6.4
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.5.2
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.19.4
packaging==20.9
pandas==1.2.0
passlib==1.7.4
pathspec==0.8.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pydantic==1.7.3
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pysaml2==6.4.1
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-http-client==3.3.1
python-multipart==0.0.5
pytz==2020.5
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.4
sendgrid==6.4.8
six==1.15.0
SQLAlchemy==1.4.0b1
SQLAlchemy-Utils==0.36.8
starkbank-ecdsa==1.1.0
starlette==0.13.6
tenacity==6.3.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.3
uvloop==0.14.0
websockets==8.1
wrapt==1.12.1
yarl==1.6.3
